multiplayer avatar updating

+++ src/network/avatar_updates.rs
@@ -2,5 +2,5 @@ use crate::renderer::transform::Transform;
 
 pub enum AvatarUpdate {
     RegisterUser(Transform, u32),
-    SetUserPosition(Transform, u32),
+    SetUserPosition(Transform, usize),
 }
